apps/api/pyproject.toml
[project]
name = "caseops-api"
version = "0.1.0"
description = "CaseOps backend API service"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
  "fastapi==0.135.2",
  "pydantic-settings==2.11.0",
  "sqlalchemy==2.0.49",
  "alembic==1.18.4",
  "uvicorn[standard]==0.43.0",
  "httpx==0.28.1",
  "email-validator==2.3.0",
  "psycopg[binary]==3.3.3",
  "PyJWT==2.12.1",
  "python-multipart==0.0.20",
  "pdfminer.six==20260107",
  "pypdfium2==5.7.0",
  "Pillow==12.2.0",
  "google-cloud-storage==3.10.1",
  "slowapi>=0.1.9",
]

[project.scripts]
caseops-document-worker = "caseops_api.workers.document_processor:main"
caseops-populate-authorities = "caseops_api.scripts.populate_authorities:main"

[dependency-groups]
dev = [
  "pytest==8.4.2",
  "pytest-asyncio==1.2.0",
  "ruff==0.13.0"
]

[build-system]
requires = ["hatchling==1.27.0"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/caseops_api"]

[tool.pytest.ini_options]
pythonpath = ["src"]
testpaths = ["tests"]

[tool.ruff]
line-length = 100
target-version = "py313"

[tool.ruff.lint]
select = ["E", "F", "I", "UP", "B"]

[tool.ruff.lint.isort]
known-first-party = ["caseops_api"]
